src/pyproject.toml
[tool.poetry]
name = "fastapi-boilerplate"
version = "0.1.0"
description = "A fully Async FastAPI boilerplate using SQLAlchemy and Pydantic 2"
authors = ["Igor Magalhaes <igor.magalhaes.r@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "fastapi_boilerplate"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
pydantic = {extras = ["email"], version = "^2.4.1"}
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
uvloop = "^0.17.0"
httptools = "^0.6.0"
uuid = "^1.30"
alembic = "^1.12.0"
asyncpg = "^0.28.0"
SQLAlchemy-Utils = "^0.41.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
SQLAlchemy = "^2.0.21"
pytest = "^7.4.2"
python-multipart = "^0.0.6"
python-decouple = "^3.8"
greenlet = "^2.0.2"
httpx = "^0.25.0"
pydantic-settings = "^2.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
